requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
asyncmy==0.2.10
pymysql==1.1.1
alembic==1.16.4
pydantic==2.11.7
pydantic-settings==2.10.1
redis==6.4.0
httpx==0.28.1
pytest==8.4.1
pytest-asyncio==1.1.0
aiosqlite==0.21.0
qwen-agent==0.0.28
prometheus-client==0.23.1
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-sqlalchemy==0.59b0
